feat: Implement unified queue execution pipeline with on-prem Kubernetes deployment
- Introduced a new `agentkernel/pipeline/` package containing the execution pipeline components: Request Handler, Input Queue, Agent Runner, Output Queue, and Response Handler.
- Added pluggable queue transports: `in_memory` (default), `sqs`, `kafka`, and `nats`.
- Implemented a pod-direct WebSocket delivery path and relocated the response store with a new `in_memory` backend.
- Created the `ak-deployment/ak-k8s/` Helm chart for Kubernetes deployment.
- Enhanced message envelope and transport interface for better queue management.
- Developed in-memory and SQS transport mechanisms, along with Kafka and NATS support.
- Generalized `AgentRunner`, `StreamAgentRunner`, `ResponseHandler`, and `RequestHandler` to work with the new pipeline architecture.
- Added new WebSocket endpoints and internal push mechanisms for real-time message delivery.
- Updated configuration management to support new transport types and response store options.
- Comprehensive testing added for new components and behaviors, ensuring compatibility with existing systems.
