+++ ak-py/src/agentkernel/pipeline/ws/__init__.py
@@ -1,0 +1,3 @@
+from .base import WebSocketConnectionStoreABC, WebSocketHandlerABC
+
+__all__ = ["WebSocketConnectionStoreABC", "WebSocketHandlerABC"]
